Add files via upload
When looking at this assignment I created three 3D objects such as a cylinder and a sphere. This assignment has the ability to move the lighting with point light. I also added rotation to the objects as well as textures that change when the mouse is clicked, a zoom feature.