Add files via upload
When looking at this file I mapped out the coordinates of both the x and y of the circle as well as the coordinates for the rectangle. This was so that the circle when hovered over covered and changed the background and when clicking on the rectangle the color of the moving circle changed.